refactor(benchmarks): simplify test suite
