+++ benchmarks/NestedChain.ts
@@ -1,17 +1,8 @@
-import {Suite} from 'benchmark'
 import * as Fluture from 'fluture'
 
-import {noop} from '../src/internals/Noop'
 import {FIO} from '../src/main/FIO'
-import {defaultRuntime} from '../src/runtimes/DefaultRuntime'
 
-import {PrintLn} from './internals/PrintLn'
-
-/**
- * Created by tushar on 2019-05-20
- */
-
-const suite = new Suite('NestedMap')
+import {RunSuite} from './internals/RunSuite'
 
 const MAX = 1e4
 
@@ -27,32 +18,4 @@ for (let i = 0; i < MAX; i++) {
   fio = fio.chain(fioMapper)
 }
 
-const runtime = defaultRuntime(undefined)
-suite
-  .add(
-    'FIO',
-    (cb: IDefer) => {
-      runtime.execute(fio, () => cb.resolve())
-    },
-    {defer: true}
-  )
-  .add(
-    'Fluture',
-    (cb: IDefer) => {
-      fluture.fork(noop, () => cb.resolve())
-    },
-    {defer: true}
-  )
-
-  .on('cycle', (event: Event) => {
-    PrintLn(String(event.target))
-  })
-  .on('complete', function(this: Suite): void {
-    PrintLn(
-      'Fastest is ' +
-        this.filter('fastest')
-          .map((i: {name: string}) => i.name)
-          .join('')
-    )
-  })
-  .run()
+RunSuite(`NestedChain ${MAX}`, {fio, fluture})
